requirements.txt
selenium
pandas
streamlit
webdriver-manager
sqlalchemy
pymysql
python-dotenv
matplotlib 
seaborn 
plotly
urllib.parse